pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "air-pollution-anomaly-detection"
version = "0.1.0"
description = "Data ingestion utilities for air quality anomaly detection."
authors = [{name = "Your Name"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
    "pandas>=2.1",
    "meteostat>=1.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
